requirements.txt
discord.py>=2.3.2,<3.0.0
python-dotenv>=1.0.1,<2.0.0
PyNaCl>=1.5.0,<2.0.0
aiohttp>=3.9.0
Pillow>=10.0.0
cryptography>=42.0.0
pyperclip>=1.8.0
pyautogui>=0.9.54
mitmproxy>=10.0.0
